requirements.txt
autopep8==1.5.4
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
cffi==1.14.3
click==7.1.2
dnspython==2.0.0
email-validator==1.1.1
Flask==1.1.2
Flask-Admin==1.5.6
Flask-BabelEx==0.9.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Security-Too==3.4.4
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
passlib==1.7.3
pycodestyle==2.6.0
pycparser==2.20
pytz==2020.1
six==1.15.0
speaklater==1.3
SQLAlchemy==1.3.19
toml==0.10.1
Werkzeug==1.0.1
WTForms==2.3.3
